Split Interaction.* on TCM-awareness
